Move sources to lib directory
