+++ .lintstagedrc
@@ -1,3 +1,3 @@
 {
-  "**/*.{js,json}": ["prettier --write", "eslint", "git add"]
+  "{test,lib}/**/*.js": ["prettier --write", "eslint", "git add"]
 }
